feat: Merge coach functionality into galaxy for unified 'Galaxy Picks'
This commit refactors the API by merging the 'coach' feature, which provides RAG-based recommendations, into the existing 'galaxy' feature. This resolves the issue of two separate, overlapping functionalities appearing in the API and Swagger UI.

Changes:
- All API endpoints from `coach.py` (for document management and recommendations) have been moved into `galaxy.py`.
- The backend worker `ai_galaxy.py`, which was a mock, has been replaced with the real RAG implementation from `ai_coach.py`.
- The `coach` router, worker, and all associated RabbitMQ configurations have been removed and files deleted.
- Frontend, documentation, and scripts have been updated to use the new unified `/galaxy/...` endpoints.
