+++ frontend/nginx.conf
@@ -10,7 +10,7 @@ server {
     }
 
     # 슬래시 유무 상관없이 프록시
-    location ~ ^/(vectorstores|files|assist|sales|events|picks)(/|$) {
+    location ~ ^/(vectorstores|files|assist|sales|events|galaxy)(/|$) {
         proxy_pass http://api:8000;
 
         proxy_set_header Host $host;
